+++ CHANGELOG.md
@@ -11,6 +11,11 @@
   Improved Tooltip widget
 - Feature [#54](https://github.com/SimformSolutionsPvtLtd/flutter_showcaseview/issues/54) - Added
   Feasibility to position tooltip left and right to the target widget.
+- Feature [#113](https://github.com/SimformSolutionsPvtLtd/flutter_showcaseview/issues/113) - Added
+  multiple showcase feature
+- Improvement [#514](https://github.com/SimformSolutionsPvtLtd/flutter_showcaseview/pull/514) -
+  Improved showcase widget and showcase with widget, Removed inherited widget, keys and setStates,
+  Added controller to manage showcase
 
 ## [4.0.1]
 - Fixed [#493](https://github.com/SimformSolutionsPvtLtd/flutter_showcaseview/issues/493) - ShowCase.withWidget not showing issue